+++ client/src/articles/Articles.jsx
@@ -1,6 +1,7 @@
 import React, { useContext, useEffect } from "react";
 import "./articles.css"; 
 
+
 const Article = ({article,lang }) => {
   
   const getLang = () => {
