+++ client/src/context/language.js
@@ -12,21 +12,21 @@ const URI = (() => {
 
 export const myContext = createContext()
 
-function ContextProvider ({ children }) {
+function ContextProvider({ children }) {
+  const URI = (() => {
+    if (process.env.NODE_ENV === "production") {
+      return "/api";
+    } else {
+      return "http://localhost:5050/api";
+    }
+  })();
   
   const [lang, setLang] = useState("he");
   const [sixTopArticles, setTopArticles] = useState();
-  const topArticles = async () => {
-    const articles = await axios.get(`${URI}`);
-    setTopArticles(articles.data);
-  };
-  useEffect(() => {
-    topArticles();
-  }, []);
 
   return (
     <myContext.Provider
-      value={{ lang, setLang, sixTopArticles, setTopArticles, topArticles}}
+      value={{ lang, setLang, sixTopArticles, setTopArticles, URI}}
     >
       {children}
     </myContext.Provider>
